separated main menu from frame; fixed look and feel menu bar.
