+++ pom.xml
@@ -6,7 +6,7 @@
 
 	<groupId>com.marginallyclever</groupId>
 	<artifactId>RobotOverlord</artifactId>
-	<version>2.112.1</version>
+	<version>2.113.0</version>
 	<name>Robot Overlord</name>
 	<description>A friendly 3D user interface for controlling robots.</description>
 	<url>https://www.marginallyclever.com/</url>
@@ -530,6 +530,7 @@
 			<version>0.3.12</version>
 		</dependency>
 
+		<!-- dockable windows -->
 		<dependency>
 			<groupId>io.github.andrewauclair</groupId>
 			<artifactId>modern-docking-api</artifactId>
@@ -545,6 +546,7 @@
 			<artifactId>modern-docking-ui</artifactId>
 			<version>0.11.1</version>
 		</dependency>
+
 		<!-- https://mvnrepository.com/artifact/com.formdev/flatlaf -->
 		<dependency>
 			<groupId>com.formdev</groupId>
