+++ src/main/java/com/marginallyclever/ro3/RO3.java
@@ -1,5 +1,7 @@
 package com.marginallyclever.ro3;
 
+import com.formdev.flatlaf.FlatLaf;
+import com.formdev.flatlaf.FlatLightLaf;
 import com.marginallyclever.ro3.apps.RO3Frame;
 
 import javax.swing.*;
@@ -10,11 +12,31 @@ import java.awt.*;
  * You can find the friendly user manual at <a href="mcr.dozuki.com">mcr.dozuki.com</a>.
  */
 public class RO3 {
+    private static final org.slf4j.Logger logger = org.slf4j.LoggerFactory.getLogger(RO3.class);
+
     public static void main(String[] args) {
         Registry.start();
+        setLookAndFeel();
 
         if(!GraphicsEnvironment.isHeadless()) {
             SwingUtilities.invokeLater(() -> (new RO3Frame()).setVisible(true));
+        }
+    }
+
+    private static void setLookAndFeel() {
+        logger.info("Setting look and feel...");
+        FlatLaf.registerCustomDefaultsSource( "com.marginallyclever.ro3" );
+        //FlatLaf.registerCustomDefaultsSource("docking");
+        try {
+            UIManager.setLookAndFeel(new FlatLightLaf());
+            // option 2: UIManager.setLookAndFeel(new FlatDarkLaf());
+        } catch (Exception ignored) {
+            logger.warn("failed to set flat look and feel. falling back to default native look and feel");
+            try {
+                UIManager.setLookAndFeel(UIManager.getSystemLookAndFeelClassName());
+            } catch (Exception ex) {
+                logger.warn("failed to set native look and feel.", ex);
+            }
         }
     }
 }
