+++ src/main/resources/com/marginallyclever/ro3/FlatLightLaf.properties
@@ -1,0 +1,14 @@
+Component.arrowType = triangle
+ScrollBar.showButtons = false
+ScrollBar.thumbInsets = 2,2,2,2
+ScrollBar.thumbArc = 999
+TabbedPane.selectedBackground = #fff
+ScrollBar.width = 14
+Panel.arc = 10
+Button.arc = 10
+Button.margin = 3, 4, 3, 4
+Component.arc = 10
+TextComponent.arc = 10
+ProgressBar.arc = 10
+Component.focusWidth = 2
+defaultFont = 12
